Relocate provider Inspect VMs button to VM table toolbar
- Move inspect button from ProviderPageHeader to TreeToolbar via
  toolbarActions slot, next to ManageColumns
- Pass provider prop to VsphereFolderTreeTable for useCanInspectProvider
- Only show button in provider details mode (!canSelect), hidden in wizard
- Extract column enrichment to module-level constant
- Remove unused useForkliftTranslation from ProviderPageHeader


Resolves: MTV-2482
