+++ src/providers/details/ProviderPageHeader.tsx
@@ -1,8 +1,5 @@
 import type { FC } from 'react';
 import { PageHeadings } from 'src/components/DetailPageHeadings/PageHeadings';
-import InspectVirtualMachinesModal, {
-  type InspectVirtualMachinesModalProps,
-} from 'src/components/InspectVirtualMachines/InspectVirtualMachinesModal';
 import LearningExperienceButton from 'src/onlineHelp/learningExperienceDrawer/LearningExperienceButton';
 import ProviderActionsDropdown from 'src/providers/actions/ProviderActionsDropdown';
 import type { ProviderData } from 'src/providers/utils/types/ProviderData';
@@ -10,21 +7,15 @@ import useGetDeleteAndEditAccessReview from 'src/utils/hooks/useGetDeleteAndEdit
 import useProviderInventory from 'src/utils/hooks/useProviderInventory';
 
 import { type ProviderInventory, ProviderModel } from '@forklift-ui/types';
-import { useModal } from '@openshift-console/dynamic-plugin-sdk';
-import { Button, ButtonVariant, Flex, FlexItem, Tooltip } from '@patternfly/react-core';
-import { TOOLTIP_TRIGGER_MANUAL } from '@utils/constants';
-import { useForkliftTranslation } from '@utils/i18n';
+import { Flex, FlexItem } from '@patternfly/react-core';
 
 import CreatePlanAction from './components/CreatePlanAction';
 import ProviderPageHeaderAlerts from './components/ProviderPageHeaderAlerts';
-import { useCanInspectProvider } from './hooks/useCanInspectProvider';
 import { useProvider } from './hooks/useProvider';
 import type { ProviderDetailsPageProps } from './utils/types';
 
 const ProviderPageHeader: FC<ProviderDetailsPageProps> = ({ name, namespace }) => {
   const { provider } = useProvider(name, namespace);
-  const { t } = useForkliftTranslation();
-  const launcher = useModal();
   const {
     error: inventoryError,
     inventory,
@@ -34,32 +25,6 @@ const ProviderPageHeader: FC<ProviderDetailsPageProps> = ({ name, namespace }) =
   });
   const permissions = useGetDeleteAndEditAccessReview({ model: ProviderModel, namespace });
   const data: ProviderData = { inventory: inventory ?? undefined, permissions, provider };
-
-  const { canInspect, disabledReason, isVsphere } = useCanInspectProvider(provider);
-
-  const onClickInspectVms = (): void => {
-    launcher<InspectVirtualMachinesModalProps>(InspectVirtualMachinesModal, {
-      provider,
-    });
-  };
-
-  const inspectButton = isVsphere ? (
-    <FlexItem>
-      <Tooltip
-        content={disabledReason}
-        trigger={disabledReason ? undefined : TOOLTIP_TRIGGER_MANUAL}
-      >
-        <Button
-          variant={ButtonVariant.secondary}
-          isDisabled={!canInspect}
-          onClick={onClickInspectVms}
-          data-testid="provider-inspect-vms-button"
-        >
-          {t('Inspect VMs')}
-        </Button>
-      </Tooltip>
-    </FlexItem>
-  ) : null;
 
   return (
     <PageHeadings
@@ -76,7 +41,6 @@ const ProviderPageHeader: FC<ProviderDetailsPageProps> = ({ name, namespace }) =
           <FlexItem>
             <LearningExperienceButton />
           </FlexItem>
-          {inspectButton}
           <FlexItem>
             <CreatePlanAction namespace={namespace} provider={provider} />
           </FlexItem>
