+++ src/utils/crds/conversion/constants.ts
@@ -60,3 +60,7 @@ export const INSPECTION_STATUS_FILTER_VALUES = [
   INSPECTION_STATUS.FAILED,
   INSPECTION_STATUS.CANCELED,
 ] as const;
+
+export const CONCURRENCY_LIMIT = 10;
+
+export const INSPECTION_TABLE_COLUMN_COUNT = 5;
